backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-environ>=0.11.2
psycopg2-binary>=2.9.9
djangorestframework-simplejwt>=5.3.0
drf-spectacular>=0.26.5
django-filter>=23.5
Pillow>=10.0
